Add more methods to the common base class.  Add CqlPreprocessorVisitor to the new inheritance hierarchy.  More cleanup and renaming.
